Initial commit of partially ported game
The current status of the port is that the game compiles and
can run in an emulator.  However, almost all of the graphics are
corrupted and everything is scaled too large for the Gamebuino's
small screen.  In addition, the hex is actually too large for a
real device and will need to be shrunk before it can actually work.

I have edited the title screen to almost fit on an 84x48 screen,
and I have fixed one of the spaceship's images to be in a format
compatible with the Gamebuino libraries.