Fixed bitmap converter, converted spaceship images
The bitmap converter wasn't yet fully functional, but now it
will convert from the strange encoding that the fast arduboy
drawBitmap routine uses.  It may not work on images that are not
powers of two.  Interestingly, examples of images with unusual
dimensions in this project all seem to not be encoded unusually
and therefore may not need to be modified.

Ship images have been converted, but the rest of the images have
yet to be fixed.  At first I planned to make the script
automatically parse and fix the image file, but after finding that
not all the images needed to be converted I doubt it's worth it
to try to fully automate the conversion process.