Fixed "phantom planet" bug and ship rotation bug
Certain planets simply could not be hailed even if they had never
been contacted before.  This appeared to be because in order to be
drawn the planet only had to be in range for x or y, not both.
I think this worked in the original game because the drawBitmap
function for the arduboy takes 16-bit coordinates, while the
gamebuino drawBitmap function only accepts 8-bit coordinates.
This resulted in planets being drawn that would normally be on
the other side of the sector.  It was a very difficult bug to
figure out.

The ship seemed to glitch whenever it was facing exactly 360
degrees, so I modified the rotation code to disallow this by
converting 360 to 0.